Calculate the table width more precisely
The `totalWidth` of the table is calculated by the width of all table header columns,
however this result calculation does not take in account when these table headers
have padding and borders.

This changes the code to calculate the total width of the table based on the
`<th>` elements vs. strings in columns, allowing use to use the `offsetWidth`
property that includes the element's padding and border in its width output.